* add default to group models  * move manager to group models  * make manager use default if it is defined and no value is given
