+++ config/local_settings.py.example
@@ -27,9 +27,10 @@ MEDIA_URL = '/media/'
 
 # Groups
 # Set the default theme for new groups: GENERAL, FOODSAVING, BIKEKITCHEN
-GROUP_THEME_DEFAULT = grouptypes.GroupTheme.GENERAL
+GROUP_THEME_DEFAULT = themes.GroupTheme.FOODSAVING
 # Set the default status for new groups, ACTIVE, PLAYGROUND, INACTIVE
-GROUP_STATUS_DEFAULT = grouptypes.GroupStatus.ACTIVE
+GROUP_STATUS_DEFAULT = themes.GroupStatus.ACTIVE
+
 
 ######
 # Mail configuration
